Introduce cx-wrapper wrapping handler function without using require/import hooks
