+++ .github/workflows/publish-nodejs.yml
@@ -25,7 +25,7 @@ jobs:
       with:
         path: ${{ env.OPENTELEMETRY_JS_CONTRIB_PATH }}
         repository: coralogix/opentelemetry-js-contrib
-        ref: refs/heads/coralogix-autoinstrumentation
+        ref: refs/heads/nodejs-cx-wrapper
         ssh-key: ${{ secrets.OPENTELEMETRY_CI_GITHUB_KEY }}
     - uses: actions/checkout@v3
       with:
