+++ nodejs/packages/cx-wrapper/.gitignore
@@ -1,0 +1,11 @@
+*.map
+common.d.ts
+common.js
+index.d.ts
+index.js
+instrumentation-init.d.ts
+instrumentation-init.js
+lambda-instrumentation-init.d.ts
+lambda-instrumentation-init.js
+provider-init.d.ts
+provider-init.js
