+++ nodejs/packages/cx-wrapper/lambda-instrumentation-init.ts
@@ -1,0 +1,93 @@
+import {
+  context as otelContext,
+  defaultTextMapGetter,
+  Context as OtelContext,
+  propagation,
+  trace,
+  diag
+} from '@opentelemetry/api';
+import { registerInstrumentations } from '@opentelemetry/instrumentation';
+import { AwsLambdaInstrumentation, AwsLambdaInstrumentationConfig } from '@opentelemetry/instrumentation-aws-lambda';
+import { OTEL_PAYLOAD_SIZE_LIMIT, OtelAttributes } from './common.js';
+
+declare global {
+  function configureLambdaInstrumentation(config: AwsLambdaInstrumentationConfig): AwsLambdaInstrumentationConfig
+}
+
+export function makeLambdaInstrumentation(): AwsLambdaInstrumentation {
+  diag.debug('Preparing handler function instrumentation');
+
+  const lambdaAutoInstrumentConfig: AwsLambdaInstrumentationConfig = {
+    requestHook: (span, { event }) => {
+      const data =
+        event && typeof event === 'object'
+          ? JSON.stringify(event)
+          : event?.toString();
+      if (data !== undefined) {
+        span.setAttribute(
+          OtelAttributes.RPC_REQUEST_PAYLOAD,
+          data.substring(0, OTEL_PAYLOAD_SIZE_LIMIT)
+        );
+      }
+    },
+    disableAwsContextPropagation: true,
+    eventContextExtractor: (event, context) => {
+      // try to extract propagation from http headers first
+      const httpHeaders = event?.headers || {};
+      const extractedHttpContext: OtelContext = propagation.extract(
+        otelContext.active(),
+        httpHeaders,
+        defaultTextMapGetter
+      );
+      if (trace.getSpan(extractedHttpContext)?.spanContext()) {
+        return extractedHttpContext;
+      }
+
+      // extract from client context
+      if (context.clientContext?.Custom) {
+        try {
+          const extractedClientContextOtelContext: OtelContext =
+            propagation.extract(
+              otelContext.active(),
+              context.clientContext.Custom,
+              defaultTextMapGetter
+            );
+          if (trace.getSpan(extractedClientContextOtelContext)?.spanContext()) {
+            return extractedClientContextOtelContext;
+          }
+        } catch (e) {
+          diag.debug(
+            'error extracting context from lambda client context payload',
+            e
+          );
+        }
+      } else if ((context.clientContext as any)?.custom) {
+        try {
+          const extractedClientContextOtelContext: OtelContext =
+            propagation.extract(
+              otelContext.active(),
+              (context.clientContext as any).custom,
+              defaultTextMapGetter
+            );
+          if (trace.getSpan(extractedClientContextOtelContext)?.spanContext()) {
+            return extractedClientContextOtelContext;
+          }
+        } catch (e) {
+          diag.debug(
+            'error extracting context from lambda client context payload',
+            e
+          );
+        }
+      }
+      return otelContext.active();
+    },
+    payloadSizeLimit: OTEL_PAYLOAD_SIZE_LIMIT,
+  };
+
+  // TODO consider not treating it as an instrumentation
+  const instrumentation = new AwsLambdaInstrumentation(typeof configureLambdaInstrumentation === 'function' ? configureLambdaInstrumentation(lambdaAutoInstrumentConfig) : lambdaAutoInstrumentConfig)
+
+  registerInstrumentations({instrumentations: [instrumentation]})
+
+  return instrumentation
+}
