+++ nodejs/packages/cx-wrapper/loader.d.ts
@@ -1,0 +1,3 @@
+import {Handler} from "aws-lambda/handler.js";
+
+export function load(taskRoot?: string, originalHandler?: string): Promise<Handler>;
