+++ nodejs/packages/cx-wrapper/loader.js
@@ -1,0 +1,3 @@
+const { load } = require('UserFunction.js'); // https://github.com/aws/aws-lambda-nodejs-runtime-interface-client/blob/main/src/UserFunction.js
+
+module.exports.load = load
